+++ apps/test-suite/tests/FileSystemNext.ts
@@ -1,6 +1,7 @@
 'use strict';
 import * as FS from 'expo-file-system';
 import { File, Directory } from 'expo-file-system/next';
+import { Paths } from 'expo-file-system/src/next';
 import { Platform } from 'react-native';
 
 export const name = 'FileSystem@next';
@@ -21,31 +22,31 @@ export async function test({ describe, expect, it, ...t }) {
   describe('FileSystem (Next)', () => {
     it('Creates a lazy file reference', () => {
       const file = new File('file:///path/to/file');
-      expect(file.path).toBe('file:///path/to/file');
+      expect(file.uri).toBe('file:///path/to/file');
     });
 
     it('Supports different slash combinations', async () => {
-      expect(new File('file:/path/to/file').path).toBe('file:///path/to/file');
+      expect(new File('file:/path/to/file').uri).toBe('file:///path/to/file');
       // FirstDirectory is a host when url parsing.
-      expect(new File('file://firstDirectory/to/file').path).toBe('file:///to/file');
-      expect(new File('file:/path/to/file').path).toBe('file:///path/to/file');
+      expect(new File('file://firstDirectory/to/file').uri).toBe('file:///to/file');
+      expect(new File('file:/path/to/file').uri).toBe('file:///path/to/file');
     });
 
-    it('Accepts and correctly handles paths to files', () => {
-      expect(new File('file:/path/to/file').path).toBe('file:///path/to/file');
-      expect(new File('file:/path/to/file/').path).toBe('file:///path/to/file');
+    it('Accepts and correctly handles uris to files', () => {
+      expect(new File('file:/path/to/file').uri).toBe('file:///path/to/file');
+      expect(new File('file:/path/to/file/').uri).toBe('file:///path/to/file');
     });
 
-    it('Accepts and correctly handles paths to directories', () => {
-      expect(new Directory('file:/path/to/file').path).toBe('file:///path/to/file/');
-      expect(new Directory('file:/path/to/file/').path).toBe('file:///path/to/file/');
+    it('Accepts and correctly handles uris to directories', () => {
+      expect(new Directory('file:/path/to/file').uri).toBe('file:///path/to/file/');
+      expect(new Directory('file:/path/to/file/').uri).toBe('file:///path/to/file/');
     });
 
-    it("Doesn't allow changing the path property", () => {
+    it("Doesn't allow changing the uri property", () => {
       const file = new File('file:///path/to/file');
       expect(() => {
         // @ts-expect-error
-        file.path = 'file:///new/path';
+        file.uri = 'file:///new/path';
       }).toThrow();
     });
 
@@ -264,23 +265,90 @@ export async function test({ describe, expect, it, ...t }) {
 
     describe('Downloads files', () => {
       it('downloads a file to a target file', async () => {
-        const url = 'https://httpbin.org/image/jpeg';
+        const url = 'https://picsum.photos/id/237/200/300';
         const file = new File(testDirectory + 'image.jpeg');
         const output = await File.downloadFileAsync(url, file);
         expect(file.exists()).toBe(true);
-        expect(output.path).toBe(file.path);
+        expect(output.uri).toBe(file.uri);
       });
 
       it('downloads a file to a target directory', async () => {
-        const url = 'https://httpbin.org/image/jpeg';
+        const url = 'https://picsum.photos/id/237/200/300';
         const directory = new Directory(testDirectory);
         const output = await File.downloadFileAsync(url, directory);
 
         const file = new File(
-          testDirectory + (Platform.OS === 'android' ? 'jpeg.jpg' : 'jpeg.jpeg')
+          testDirectory + (Platform.OS === 'android' ? '300.jpg' : '237-200x300.jpg')
         );
         expect(file.exists()).toBe(true);
-        expect(output.path).toBe(file.path);
+        expect(output.uri).toBe(file.uri);
+      });
+    });
+
+    describe('Computes file properties', () => {
+      it('computes size', async () => {
+        const file = new File(testDirectory + 'file.txt');
+        file.write('Hello world');
+        expect(file.size).toBe(11);
+      });
+
+      it('computes md5', async () => {
+        const file = new File(testDirectory + 'file.txt');
+        file.write('Hello world');
+        expect(file.md5).toBe('3e25960a79dbc69b674cd4ec67a72c62');
+      });
+
+      it('returns null size and md5 for nonexistent files', async () => {
+        const file = new File(testDirectory + 'file2.txt');
+        expect(file.size).toBe(null);
+        expect(file.md5).toBe(null);
+      });
+
+      it('computes md5', async () => {
+        const file = new File(testDirectory + 'file.txt');
+        file.write('Hello world');
+        expect(file.md5).toBe('3e25960a79dbc69b674cd4ec67a72c62');
+      });
+    });
+
+    describe('Returns base64', () => {
+      it('gets base64 of a file', async () => {
+        const src = new File(testDirectory + 'file.txt');
+        src.write('Hello world');
+        expect(src.base64()).toBe('SGVsbG8gd29ybGQ=');
+      });
+    });
+
+    describe('Lists directory contents', () => {
+      it('for newly created directories', () => {
+        new File(testDirectory + 'file.txt').create();
+        new Directory(testDirectory + 'directory').create();
+        expect(new Directory(testDirectory).list()).toEqual([
+          new File(testDirectory + 'file.txt'),
+          new Directory(testDirectory + 'directory'),
+        ]);
+        expect(new Directory(testDirectory).list()[0] instanceof File).toBe(true);
+      });
+    });
+
+    describe('JS-only properties for path manipulation', () => {
+      it('return parentDirectory for files', () => {
+        const file = new File(testDirectory + 'image.jpeg');
+        expect(file.parentDirectory.uri).toBe(new Directory(testDirectory).uri);
+      });
+      it('return parentDirectory for directories', () => {
+        const directory = new Directory(testDirectory + '/testdirectory/sampleDir');
+        expect(directory.parentDirectory.parentDirectory.uri).toBe(
+          new Directory(testDirectory).uri
+        );
+      });
+      it('return extension for files', () => {
+        expect(new File(testDirectory + 'image.jpeg').extension).toBe('.jpeg');
+        expect(new File(testDirectory + 'image.pdf.jpeg').extension).toBe('.jpeg');
+      });
+
+      it('joins paths', () => {
+        expect(Paths.join('file:///path', 'to', '..', 'file')).toBe('file:///path/file');
       });
     });
   });