+++ packages/expo-file-system/android/src/main/java/expo/modules/filesystem/next/FileSystemNextModule.kt
@@ -11,6 +11,7 @@ import expo.modules.kotlin.types.Either
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import java.io.File
+import java.io.FileNotFoundException
 import java.io.FileOutputStream
 import java.net.URI
 
@@ -34,9 +35,9 @@ class FileSystemNextModule : Module() {
       val fileName = URLUtil.guessFileName(url.toString(), contentDisposition, contentType)
 
       val destination = if (to is FileSystemDirectory) {
-        File(to.path, fileName)
+        File(to.file, fileName)
       } else {
-        to.path
+        to.file
       }
 
       val body = response.body ?: throw UnableToDownloadException("response body is null")
@@ -49,8 +50,8 @@ class FileSystemNextModule : Module() {
     }
 
     Class(FileSystemFile::class) {
-      Constructor { path: URI ->
-        FileSystemFile(File(path.path))
+      Constructor { uri: URI ->
+        FileSystemFile(File(uri.path))
       }
 
       Function("delete") { file: FileSystemFile ->
@@ -81,6 +82,10 @@ class FileSystemNextModule : Module() {
         file.text()
       }
 
+      Function("base64") { file: FileSystemFile ->
+        file.base64()
+      }
+
       Function("exists") { file: FileSystemFile ->
         file.exists()
       }
@@ -93,14 +98,26 @@ class FileSystemNextModule : Module() {
         file.move(destination)
       }
 
-      Property("path") { file ->
+      Property("uri") { file ->
         file.asString()
+      }
+
+      Property("md5") { file ->
+        try {
+          file.md5
+        } catch (e: FileNotFoundException) {
+          null
+        }
+      }
+
+      Property("size") { file ->
+        file.size
       }
     }
 
     Class(FileSystemDirectory::class) {
-      Constructor { path: URI ->
-        FileSystemDirectory(File(path.path))
+      Constructor { uri: URI ->
+        FileSystemDirectory(File(uri.path))
       }
 
       Function("delete") { directory: FileSystemDirectory ->
@@ -127,8 +144,13 @@ class FileSystemNextModule : Module() {
         directory.move(destination)
       }
 
-      Property("path") { directory ->
+      Property("uri") { directory ->
         directory.asString()
+      }
+
+      // this function is internal and will be removed in the future (when returning arrays of shared objects is supported)
+      Function("listAsRecords") { directory: FileSystemDirectory ->
+        directory.listAsRecords()
       }
     }
   }