+++ packages/expo-modules-core/ios/AppDelegates/ExpoAppDelegate.swift
@@ -301,7 +301,7 @@ open class ExpoAppDelegate: UIResponder, UIApplicationDelegate {
    * a different orientation.
    */
 #if !os(tvOS)
-  public func application(_ application: UIApplication, supportedInterfaceOrientationsFor window: UIWindow?) -> UIInterfaceOrientationMask {
+  open func application(_ application: UIApplication, supportedInterfaceOrientationsFor window: UIWindow?) -> UIInterfaceOrientationMask {
     let deviceOrientationMask = allowedOrientations(for: UIDevice.current.userInterfaceIdiom)
     let universalOrientationMask = allowedOrientations(for: .unspecified)
     let infoPlistOrientations = deviceOrientationMask.isEmpty ? universalOrientationMask : deviceOrientationMask