+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoModule.kt
@@ -2,7 +2,6 @@
 
 package expo.modules.video
 
-import android.app.Activity
 import android.net.Uri
 import androidx.media3.common.C
 import androidx.media3.common.PlaybackParameters
@@ -14,7 +13,6 @@ import com.facebook.react.uimanager.Spacing
 import com.facebook.react.uimanager.ViewProps
 import com.facebook.yoga.YogaConstants
 import expo.modules.kotlin.apifeatures.EitherType
-import expo.modules.kotlin.exception.Exceptions
 import expo.modules.kotlin.modules.Module
 import expo.modules.kotlin.modules.ModuleDefinition
 import expo.modules.kotlin.types.Either
@@ -28,9 +26,6 @@ import kotlinx.coroutines.runBlocking
 // https://developer.android.com/guide/topics/media/media3/getting-started/migration-guide#improvements_in_media3
 @androidx.annotation.OptIn(androidx.media3.common.util.UnstableApi::class)
 class VideoModule : Module() {
-  private val activity: Activity
-    get() = appContext.activityProvider?.currentActivity ?: throw Exceptions.MissingActivity()
-
   override fun definition() = ModuleDefinition {
     Name("ExpoVideo")
 
@@ -39,7 +34,7 @@ class VideoModule : Module() {
     }
 
     Function("isPictureInPictureSupported") {
-      return@Function VideoView.isPictureInPictureSupported(activity)
+      return@Function VideoView.isPictureInPictureSupported(appContext.throwingActivity)
     }
 
     View(VideoView::class) {
@@ -152,7 +147,7 @@ class VideoModule : Module() {
 
     Class(VideoPlayer::class) {
       Constructor { source: VideoSource? ->
-        val player = VideoPlayer(activity.applicationContext, appContext, source)
+        val player = VideoPlayer(appContext.throwingActivity.applicationContext, appContext, source)
         appContext.mainQueue.launch {
           player.prepare()
         }